add MESSAGE_OPTIONS enum to ConsentLib
When the user clicks on an "option" inside the message, the javascript returns a integer. By mapping this integer into one of the values inside MESSAGE_OPTIONS we make it easier for the user to understand what the options mean.
At the moment, there're only two options mapped:
12 -> SHOW_PRIVACY_MANAGER
Any -> UNKNOWN